requirements.txt
django==1.11.7
djangorestframework==3.7.3
django-cors-headers==2.1.0
